==> 00_notas.ipynb <==
https://github.com/Rossjomo/proyecto_final/tree/main

Los datos: Estos datos son de un censo de EE. UU. Del Archivo Train.csv https://www.kaggle.com/datasets/kamaumunyori/income-prediction-dataset-us-20th-century-data/data

¿Gana más de 50 000 dólares? Este número es significativo porque en 2022 "el ingreso medio fue de 40 480 dólares y el ingreso medio fue de 59 430 dólares" (1)

Esta cifra ha cambiado en los últimos años desde el promedio hasta el límite inferior de la clase media y se considera que "50 000 dólares pueden ser un buen salario de nivel inicial" 
Queríamos descubrir qué impacto tendría ser un migrante en los EE. UU. al ganar más de ese umbral de 50 000 dólares.

(2)

(1) https://en.wikipedia.org/wiki/Personal_income_in_the_United_S
tates (2) https://www.sofi.com/learn/content/is-50k-a-good-salary/

Notas: msa - Metropolitan statistial area

La edad legal para trabajar en EE. UU. es de 14 años

# SQL connection

import sqlite3
import pandas as pd

# Read the CSV file into a pandas DataFrame
df = pd.read_csv('Train.csv')

# Connect to the SQLite database (or create it if it doesn't exist)
conn = sqlite3.connect('local_database.db')

# Create the table (optional, if the table does not already exist)
# Make sure to define the table structure based on your CSV columns.
df.to_sql('my_table', conn, if_exists='replace', index=False)


query_below = "SELECT * FROM my_table WHERE income_above_limit = 'Below limit'"
result_below = pd.read_sql(query_below, conn)

query_above = "SELECT * FROM my_table WHERE income_above_limit = 'Above limit'"
result_above = pd.read_sql(query_above, conn)

# Export to CSV
result_below.to_csv('sql_query_below.csv', index=False)

result_above.to_csv('sql_query_above.csv', index=False)

# Close the connection
conn.close()
