==> 02_limpiar.ipynb <==
# Importar Librarias y datos

import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import numpy as np   
import itertools

from scipy.stats import chi2_contingency, ttest_ind
from statsmodels.stats.outliers_influence import variance_inflation_factor
import statsmodels.formula.api as smf 

df = pd.read_parquet("output_01.parquet")

df.head()

# Corregir nombres de columnas

# Cambiar el nombre de una columna en especifico
df.rename(columns={'mig_year': 'migration_year'}, inplace=True)

df.info()

# Corregir valores categóricos

# Identificar columnas categóricas
columnas_categoricas = df.select_dtypes(include=['object']).columns

columnas_categoricas

# Al revisar los valores vemos que tienen espacios en blanco
for col in columnas_categoricas:
    print(f"Valores únicos en '{col}': {df[col].unique()}")

# Aplicar correcciones a todas las columnas categóricas quitando todos los espacios en blanco
df[columnas_categoricas] = df[columnas_categoricas].apply(
    lambda col: col.str.strip()
)

# Volver a revisar valores únicos de todas las columnas categóricas
for col in columnas_categoricas:
    print(f"Valores únicos en '{col}': {df[col].unique()}")

# Buscar filas duplicadas

duplicado = df[df.duplicated()]
duplicado

No hay filas duplicadas entonces no hay instances para remover

# Buscar nulos

df.isnull().sum()


# Crear una copia del dataset combinado
df_copia = df.copy()

# Lista de columnas a eliminar
columnas_a_eliminar = [
    'ID', 
    'is_labor_union', 
    'industry_code', 
    'occupation_code_main', 
    'household_summary',
    'under_18_family', 
    'veterans_admin_questionnaire', 
    'vet_benefit',
    'mig_year', 
    'country_of_birth_father', 
    'country_of_birth_mother', 
    'migration_prev_sunbelt', 
    'migration_code_move_within_reg',
    'residence_1_year_ago',
    'old_residence_reg',
    'old_residence_state', 
    'importance_of_record'
]

# Eliminar las columnas en la copia
df_copia = df_copia.drop(columns=columnas_a_eliminar, errors='ignore')

# Mostrar las primeras filas del dataset modificado
df_copia.head()


# Filtrar columnas que tiene mas que 40% nulos

columnas_reducidas = df_copia.loc[:, (df_copia.isna().sum() / len(df_copia) * 100) < 40]
print(columnas_reducidas.shape)
print(columnas_reducidas.columns)

columnas_reducidas.shape

# Arreglar los nulos

De las columnas restantes, busque valores nulos. Los únicos valores nulos están en las columnas categóricas, por lo que se rellenan con la moda de la columna.

columnas_reducidas.isna().sum()

# ver values y el mas común 
print(columnas_reducidas['migration_code_change_in_msa'].unique())
print("Value mas comun:", columnas_reducidas['migration_code_change_in_msa'].mode()[0])

print(columnas_reducidas['migration_code_change_in_reg'].unique())
print("Value mas comun:", columnas_reducidas['migration_code_change_in_reg'].mode()[0])

columns_to_process = [
    'migration_code_change_in_msa',
    'migration_code_change_in_reg'
]

for column in columns_to_process:
    if column in columnas_reducidas:
        columnas_reducidas.loc[:, column] = (
            columnas_reducidas[column]
            .replace({' ?': 'Unknown'})
            .fillna('Unknown')
        )


# verificar los valores
columnas_reducidas.isna().sum()

columnas_reducidas.shape

# Guardar output

columnas_reducidas.to_parquet("output_02.parquet")
