==> 04_standardizar.ipynb <==
# Importar Librarias y datos

import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import numpy as np   
import itertools

from scipy.stats import chi2_contingency, ttest_ind
from statsmodels.stats.outliers_influence import variance_inflation_factor
import statsmodels.formula.api as smf 

df = pd.read_parquet("output_02.parquet") # no es 03 porque visualisar no ha afectado

df.head()

# Modificar datos a binario

df.info()

df['country_of_birth_own'].unique()

# change US to US & Other
df.loc[df['country_of_birth_own'] != 'US', 'country_of_birth_own'] = 'Other'
df['country_of_birth_own'].unique()

df['gender'].unique()

df['gender'] = df['gender'].replace({'Male':0, 'Female':1})
df['gender'].unique()

df['income_above_limit'] = df['income_above_limit'].replace({'Above limit':1, 'Below limit':0})
df['income_above_limit'].unique()

# limpieza de valores categoricas 

Reducir los valores en los datos categóricos para eliminar la repetición y facilitar la manipulación de datos.

df['household_stat'].unique()

df['education'].unique()

df_copia = df.copy()

mapping = {
    'High school graduate': 'secondary_complete', 
    '12th grade no diploma': 'secondary_incomplete', 
    'Children': 'child',
    'Bachelors degree(BA AB BS)': 'university_graduate_or_higher', 
    '7th and 8th grade': 'secondary_incomplete', 
    '11th grade': 'secondary_incomplete',
    '9th grade':'secondary_incomplete', 
    'Masters degree(MA MS MEng MEd MSW MBA)': 'university_graduate_or_higher',
    '10th grade': 'secondary_incomplete', 
    'Associates degree-academic program': 'technical_education/university_incomplete',
    '1st 2nd 3rd or 4th grade': 'primary', 
    'Some college but no degree': 'technical_education/university_incomplete',
    'Less than 1st grade': 'primary', 
    'Associates degree-occup /vocational': 'technical_education/university_incomplete',
    'Prof school degree (MD DDS DVM LLB JD)':'university_graduate_or_higher', 
    '5th or 6th grade': 'primary',
}

df_copia['education'] = df_copia['education'].replace(mapping)

df_copia['education'].unique()

mapping = {
    'Householder': 'Householder',
    'Nonfamily householder': 'Nonfamily Householder',
    'Spouse of householder': 'Spouse',
    'Child 18+ never marr Not in a subfamily': 'Child',
    'Child <18 never marr not in subfamily': 'Child',
    'Other Rel 18+ spouse of subfamily RP': 'Other Relative',
    'In group quarters': 'In Group Quarters',
    'Child 18+ spouse of subfamily RP': 'Child', 
    'Secondary individual': 'Other',
    'Child 18+ never marr RP of subfamily': 'Child',
    'Other Rel 18+ spouse of subfamily RP': 'Other',
    'Grandchild <18 never marr not in subfamily': 'Grandchild',
    'Other Rel <18 never marr child of subfamily RP': 'Other',
    'Other Rel 18+ ever marr RP of subfamily': 'Other',
    'Other Rel 18+ ever marr not in subfamily': 'Other',
    'Child 18+ ever marr Not in a subfamily': 'Child',
    'RP of unrelated subfamily': 'Other',
    'Child 18+ ever marr RP of subfamily': 'Child',
    'Other Rel 18+ never marr not in subfamily': 'Other',
    'Child under 18 of RP of unrel subfamily': 'Child',
    'Grandchild <18 never marr child of subfamily RP': 'Grandchild',
    'Grandchild 18+ never marr not in subfamily': 'Grandchild',
    'Other Rel <18 never marr not in subfamily': 'Other', 
    'In group quarters': 'Group Quarters',
    'Grandchild 18+ ever marr not in subfamily': 'Grandchild',
    'Other Rel 18+ never marr RP of subfamily': 'Other',
    'Child <18 never marr RP of subfamily': 'Child',
    'Grandchild 18+ never marr RP of subfamily': 'Grandchild',
    'Spouse of RP of unrelated subfamily': 'Spouse',
    'Grandchild 18+ ever marr RP of subfamily': 'Grandchild',
    'Child <18 ever marr not in subfamily': 'Child',
    'Child <18 ever marr RP of subfamily': 'Child',
    'Other Rel <18 ever marr RP of subfamily': 'Other',
    'Grandchild 18+ spouse of subfamily RP': 'Grandchild',
    'Child <18 spouse of subfamily RP': 'Child',
    'Other Rel <18 ever marr not in subfamily': 'Other',
    'Other Rel <18 never married RP of subfamily': 'Other',
    'Other Rel <18 spouse of subfamily RP': 'Other',
    'Grandchild <18 ever marr not in subfamily': 'Grandchild',
    'Grandchild <18 never marr RP of subfamily': 'Grandchild'
}
df_copia['household_stat'] = df_copia['household_stat'].map(mapping).fillna('Other')

df_copia['household_stat'].unique()

# Guardar output

df_copia.to_parquet("output_04.parquet")
