==> 05_P_valor_Vif.ipynb <==
# Importar Librarias y datos

import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import numpy as np   
import itertools

from scipy.stats import chi2_contingency, ttest_ind
from statsmodels.stats.outliers_influence import variance_inflation_factor
import statsmodels.formula.api as smf 

df_train = pd.read_parquet("output_04.parquet")

df_train.head()

# **P-Valor** 

print("\nEstadísticas descriptivas:")
print(df_train.describe())

###Este bloque de código genera las estadísticas descriptivas del conjunto de datos, tales como media, mediana, desviación estándar, y valores mínimos y máximos de las columnas numéricas.

##Es útil para comprender la dispersión y distribución de los datos.

# Identificación de columnas numéricas y categóricas

En este paso se identifican las columnas numéricas y categóricas dentro del conjunto de datos. Esto es importante para aplicar correctamente los métodos de preprocesamiento, ya que los datos numéricos y categóricos requieren tratamientos diferentes.


# Identificamos las columnas numéricas y categóricas
numeric_cols = ['age', 'wage_per_hour', 'working_week_per_year', 'total_employed', 'gains', 'losses', 'income_above_limit']
categorical_cols = ['gender', 'education', 'marital_status', 'race', 'is_hispanic', 
                    'employment_commitment', 'employment_stat', 'industry_code_main', 
                    'occupation_code', 'household_stat', 'tax_status', 'stocks_status', 
                    'citizenship', 'country_of_birth_own', 'migration_code_change_in_as', 
                    'migration_code_change_in_reg']


# Transformación de columnas numéricas y categoricas

En este paso se implementan dos transformaciones diferentes para las columnas numéricas y categóricas utilizando pipelines de scikit-learn. Estas transformaciones se aplicarán a los datos antes de entrenar el modelo.

Para las columnas numéricas y categoricas, se define un pipeline con los siguientes pasos:

Imputer (SimpleImputer): Rellena los valores faltantes utilizando la media de cada columna numérica.

Scaler (StandardScaler): Estandariza las variables numéricas transformando sus valores para que tengan una media de 0 y una desviación estándar de 1. Esto es importante para garantizar que todas las características numéricas estén en la misma escala y que no dominen aquellas con valores más altos.

Imputer (SimpleImputer): Rellena los valores faltantes utilizando la moda (el valor más frecuente) en cada columna categórica.

OneHotEncoder: Convierte las variables categóricas en variables dummy (0 y 1) utilizando la codificación One-Hot, eliminando la primera categoría con drop='first' para evitar la multicolinealidad

# Importar librerías necesarias
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.impute import SimpleImputer
from sklearn.compose import ColumnTransformer
import matplotlib.pyplot as plt
import seaborn as sns  



# Definir columnas numéricas y categóricas

# Convertir income_above_limit a numérico antes del preprocesamiento
#df_train['income_above_limit'] = df_train['income_above_limit'].map({'Yes': 1, 'No': 0})


numeric_cols = ['age', 'wage_per_hour', 'working_week_per_year', 'total_employed', 'gains', 'losses', 'income_above_limit']
categorical_cols = ['gender', 'education', 'marital_status', 'race', 'is_hispanic', 
                    'employment_commitment', 'employment_stat', 'industry_code_main', 
                    'occupation_code', 'household_stat', 'tax_status', 'stocks_status', 
                    'citizenship', 'country_of_birth_own', 'migration_code_change_in_as', 
                    'migration_code_change_in_reg',]

# Preprocesamiento para columnas numéricas
numeric_transformer = Pipeline(steps=[
    ('imputer', SimpleImputer(strategy='mean')),  # Imputación con la media
    ('scaler', StandardScaler())  # Escalado estándar
])

# Preprocesamiento para columnas categóricas
categorical_transformer = Pipeline(steps=[
    ('imputer', SimpleImputer(strategy='most_frequent')),  # Imputación con el valor más frecuente
    ('onehot', OneHotEncoder(drop='first'))  # Codificación one-hot
])

# Transformador de columnas
preprocessor = ColumnTransformer(
    transformers=[
        ('num', numeric_transformer, numeric_cols),
        ('cat', categorical_transformer, categorical_cols)
    ])

# Confirmación de cambios
print("Lo Logramos :).")


# Ver los valores únicos de la variable
print(df_train['income_above_limit'].unique())

"""
# Mapear valores a números
df_train['income_above_limit'] = df_train['income_above_limit'].map({
    'Below limit': 0,
    'Above limit': 1
})
"""
# Verificar que la conversión fue exitosa
print(df_train['income_above_limit'].dtype)
print(df_train['income_above_limit'].value_counts())

# Eliminar posibles valores NaN resultantes de la conversión
df_train = df_train.dropna(subset=['income_above_limit'])


modelo = smf.ols(
    formula="income_above_limit ~ age + education + marital_status + employment_stat + working_week_per_year + occupation_code + is_hispanic + gender + race",
    data=df_train
).fit()

# 6. Obtener p-valores
p_values = modelo.pvalues
significant_vars = p_values[p_values < 0.05]
print(significant_vars)

col_relevantes = modelo.pvalues[modelo.pvalues<0.05]
col_relevantes

# **Conclusiones del análisis de los p-valores**

### La educación y el estado civil tienen un impacto significativo en los ingresos  
- Los coeficientes de varias categorías de educación, como **doctorado, maestría, licenciatura y secundaria**, tienen p-valores extremadamente bajos (próximos a 0).  
  Esto indica que el nivel educativo tiene una fuerte relación con la probabilidad de superar el límite de ingresos.  
- El estado civil también es relevante: estar **casado con cónyuge civil presente** o **nunca haberse casado** son factores significativos.  

### El género y la raza también influyen en los ingresos  
- **Ser hombre (`gender[T.Male]`)** tiene un p-valor de **0.000000**, lo que sugiere una diferencia significativa en ingresos entre hombres y mujeres.  
- La variable **`race[T.White]`** es significativa, aunque con un p-valor más alto (**0.0393**), indicando que la raza puede tener una relación con el nivel de ingresos,  
  pero es menos determinante que otras variables como educación o empleo.  

###  **Conclusión Final**  
La **educación**, el **estado civil**, el **género** y la **ocupación** son los principales factores asociados con ingresos elevados en este modelo. 


# **Vif**


from sklearn.linear_model import LinearRegression
from sklearn.model_selection import train_test_split

#columnas_numericas = df_heart.select_dtypes(include=['float64', 'int64']).columns
columnas = df_train.columns[df_train.dtypes != 'object']
# Convertir índices a listas
columnas = columnas.tolist()
columnas

# Analizamos variables numericas del dataset
df_continuas = df_train[columnas]
df_continuas.columns

vif_acum = {}

for i in columnas:
   
    x = df_continuas.loc[:,df_continuas.columns != i]
    y = df_continuas.loc[:,i]

    lm = LinearRegression()
    lm.fit(x, y)
    r2 = lm.score(x, y)

    vif = (1/(1-r2))
    vif_acum[i] = vif

vif_acum

## Conclusiones del Análisis de VIF

### No hay signos fuertes de multicolinealidad en la mayoría de las variables  
- La mayoría de los valores de **VIF** están por debajo de **5**, lo que indica que las variables no presentan una colinealidad significativa.  
- Variables como **age (1.06)**, **wage_per_hour (1.05)**, **gains (1.03)** y **losses (1.02)** tienen valores muy cercanos a **1**, lo que significa que prácticamente no están correlacionadas con otras variables.  
- La variable **migration_year** tiene el VIF más bajo (**1.00**), lo que indica que no está influenciada por ninguna otra variable en el conjunto de datos.  

### Posible colinealidad moderada en *working_week_per_year*, *total_employed* y *occupation_code*  
- **working_week_per_year** tiene un **VIF de 3.03**, el más alto del conjunto, lo que sugiere que puede tener cierta correlación con otras variables, aunque sigue en un rango aceptable.  
- **total_employed (2.28)** y **occupation_code (1.79)** también tienen valores algo elevados en comparación con otras variables, lo que podría indicar que comparten información con otras características del modelo.  
- Aunque estos valores no son alarmantes, si en un análisis más profundo se detecta impacto en la interpretación del modelo, se podría considerar eliminar una de estas variables o combinar información redundante.
