==> 06_revision_de_datos.ipynb <==
# Importar Librarias y datos

import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import numpy as np   
import itertools

from scipy.stats import chi2_contingency, ttest_ind
from statsmodels.stats.outliers_influence import variance_inflation_factor
import statsmodels.formula.api as smf 

df = pd.read_parquet("output_04.parquet")

df.head()

# Revisión de los datos

Los nativos no deberían tener "migration_year"

Se verificó el año y se compararon las otras columnas de migración:

- `'migration_code_change_in_msa'`
- `'migration_code_change_in_reg'`  


Se eliminó esta columna de año porque muestra que se registró la información de migración o, si no se registró información, el año se indica como `"95"` .

Parece que mig_year (1994 o 1995) está relacionado con la columna migration_code_change_in_msa. Para entenderlo mejor, desglosémoslo:

Censo de EE.UU.: Posiblemente, los datos provienen del Current Population Survey (CPS) o del Censo de EE.UU., que suelen incluir preguntas sobre migración interna.
Columna migration_code_change_in_msa:
"MSA" significa Metropolitan Statistical Area (Área Estadística Metropolitana).
Esta columna probablemente indica si una persona cambió de MSA en el último año.
Si tiene algún dato (un código), podría significar que hubo un cambio de MSA.
Si tiene "?", podría significar que no hay información disponible o que la persona no migró.

Posible conexión con mig_year:

Cambio en la forma de recopilar datos
Puede que en 1994 el censo registrara información más detallada sobre migración dentro de las MSA.
En 1995, es posible que haya habido un cambio en la encuesta, lo que resultó en valores "?" para esa columna.

Diferencias en la población encuestada:

Si mig_year = 1994 significa que la persona respondió a la encuesta en 1994, entonces en ese año las personas tenían datos en migration_code_change_in_msa.
Si mig_year = 1995, quizás la encuesta cambió y ya no se preguntó o registró de la misma manera.

df.columns

df[['citizenship', 'country_of_birth_own', 'migration_code_change_in_msa', 'migration_code_change_in_reg', 'migration_year']][(df['citizenship'] == "Native") & (df['migration_code_change_in_msa'] == '?')].head(25)

df[['citizenship', 'country_of_birth_own', 'migration_code_change_in_msa', 'migration_code_change_in_reg', 'migration_year']][(df['citizenship'] == "Native") & (df['migration_code_change_in_msa'] != '?')].head(25)

df_copia = df.drop(columns=['migration_year'])

# Filtro de edad

La edad legal para trabajar en EE. UU. es de 14 años. Filtra los resultados para incluir solo a las personas que están legalmente autorizadas a ganar un salario.

df_copia_mas_que_14 = df_copia[df_copia['age'] > 14]

df_copia_mas_que_14.head()

# Guardar output

df_copia_mas_que_14.to_parquet("output_06.parquet")
