==> 09.5_modelo_KNN_metricas (1).ipynb <==
import pandas as pd

# Load all model metrics
knn = pd.read_csv("knn_metrics.csv")
knn_ov = pd.read_csv("knn_ov.csv")
knn_un = pd.read_csv("knn_un.csv")
knn_com = pd.read_csv("knn_com.csv")
knn_thresh = pd.read_csv("knn_threshold.csv")

# Combine all into a single DataFrame
all_models_df = pd.concat([knn, knn_ov, knn_un, knn_com, knn_thresh], ignore_index=True)
 
# drop first column
models = all_models_df.iloc[:, 1:]

# save to csv
models.to_csv("model_metrics.csv")

models
